# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Anime Calendar API
  version: 1.0.0
  description: |
    REST API for Anime Calendar — create and manage personal anime/manga calendars
    sourced from AniList, exportable as iCalendar (.ics) feeds.

    ## Authentication
    Most endpoints require a Bearer JWT obtained via `/login`, `/register`, or `/auth/google`.
    Include it as: `Authorization: Bearer <token>`

    ## Error responses
    All errors return a JSON body: `{"error": "<message>"}`.
    The HTTP status code indicates the error category.
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable response components
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Requested resource not found
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid request data
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Requested resource not found
...
    Language:
      type: string
      enum: [english, native, romaji]
      description: Title language preference for AniList titles
      default: english
    Item:
      type: object
      required: [id, id_mal, title, media_type, episode_duration, airing_schedule, cover_image, banner_image]
      properties:
        id:
          type: integer
          description: AniList media ID
          example: 1535
        id_mal:
          type: integer
          nullable: true
          description: MyAnimeList ID (nullable)
          example: 16498
        title:
          type: object
          properties:
            english:
              type: string
              nullable: true
              example: Attack on Titan
            native:
              type: string
              nullable: true
              example: 進撃の巨人
            romaji:
              type: string
              nullable: true
              example: Shingeki no Kyojin
        media_type:
          type: string
          enum: [ANIME, MANGA]
        episode_duration:
          type: integer
          description: Episode duration in minutes
          example: 24
        airing_schedule:
          type: array
          items:
            $ref: '#/components/schemas/AiringSchedule'
        cover_image:
          type: object
          properties:
            medium:
              type: string
              format: uri
              nullable: true
            large:
              type: string
              format: uri
              nullable: true
            extra_large:
              type: string
              format: uri
              nullable: true
            color:
              type: string
              nullable: true
              example: '#2b2d42'
        banner_image:
          type: string
          format: uri
          nullable: true
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
    AiringSchedule:
      type: object
      required: [episode, airing_at]
      properties:
        episode:
          type: integer
          example: 1
        airing_at:
          type: integer
          description: Unix timestamp of scheduled air time
          example: 1680000000
    Recommendation:
      type: object
      required: [rating, media]
      properties:
        rating:
          type: integer
          description: Community rating for this recommendation
          example: 42
        media:
          type: object
          required: [id, id_mal, title, cover_image]
          properties:
            id:
              type: integer
            id_mal:
              type: integer
              nullable: true
            title:
              type: object
              properties:
                english:
                  type: string
                  nullable: true
                native:
                  type: string
                  nullable: true
                romaji:
                  type: string
                  nullable: true
            cover_image:
              type: object
              properties:
                medium:
                  type: string
                  format: uri
                  nullable: true
                large:
                  type: string
                  format: uri
                  nullable: true
                extra_large:
                  type: string
                  format: uri
                  nullable: true
                color:
                  type: string
                  nullable: true
    Calendar:
      type: object
      required: [id, name, language, items, created_at, updated_at]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 100
          example: My Spring 2024 Anime
        language:
          $ref: '#/components/schemas/Language'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CalendarRequest:
      type: object
      required: [name, language, items]
      properties:
        id:
          type: integer
          default: 0
          description: 0 (or omit) to create; existing ID to update
          example: 0
        name:
          type: string
          maxLength: 100
          example: My Spring 2024 Anime
        language:
          $ref: '#/components/schemas/Language'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
    PageCalendar:
      type: object
      required: [id, name, item_count, subscription_token, created_at, updated_at]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: My Spring 2024 Anime
        item_count:
          type: integer
          description: Number of items in the calendar
          example: 12
        subscription_token:
          type: string
          description: Stable token for the iCal subscription URL
          example: aB3xZ9...
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaginatedCalendars:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PageCalendar'
        pagination:
          type: object
          required: [page, page_size, total, total_pages]
          properties:
            page:
              type: integer
              example: 1
            page_size:
              type: integer
              example: 6
            total:
              type: integer
              description: Total number of calendars
              example: 14
            total_pages:
              type: integer
              example: 3
...
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /calendars:
    get:
      tags: [Calendars]
      summary: List the authenticated user's calendars (paginated)
      operationId: listCalendars
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 6
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated list of calendars
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCalendars'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /calendar:
    put:
      tags: [Calendars]
      summary: Create or update a calendar
      operationId: upsertCalendar
      description: |
        Upserts a calendar. Set `id` to `0` (or omit it) to create a new calendar.
        Provide an existing `id` to update. The item list fully replaces the current items.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarRequest'
      responses:
        '200':
          description: Calendar created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /calendars/{id}:
    get:
      tags: [Calendars]
      summary: Get a single calendar with its full items
      operationId: getCalendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Calendar with items populated from AniList
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Calendars]
      summary: Delete a calendar
      operationId: deleteCalendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Calendar deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /calendars/{id}/export:
    get:
      tags: [Calendars]
      summary: Download the calendar as an .ics file
      operationId: exportCalendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: iCalendar file
          content:
            text/calendar:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /calendars/subscribe/{token}:
    get:
      tags: [Calendars]
      summary: Public iCal subscription feed
      operationId: subscribeCalendar
      description: |
        Returns the calendar as an iCal feed. This endpoint is unauthenticated —
        the `subscription_token` acts as the access credential. Suitable for use
        as a subscription URL in Google Calendar, Apple Calendar, etc.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: The `subscription_token` from the calendar object
      responses:
        '200':
          description: iCalendar feed
          content:
            text/calendar:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------- Items --------------------------------------------------------
...